[modify] 환경변수 치환 함수 재배치
